Enelver les view useless
