+++ frontend/src/router/index.ts
@@ -8,17 +8,14 @@ import MenuManagement from '../views/MenuManagement.vue';
 import OrderManagement from '../views/OrderManagement.vue';
 import StatisticsManagement from '../views/StatisticsManagement.vue';
 import UserView from '../views/UserView.vue';
-import ReservationView from '../views/ReservationView.vue';
-import ReservationChoiceView from '../views/ReservationChoiceView.vue';
 import FeedbackView from '../views/FeedbackView.vue';
 import ModifyView from '../views/ModifyView.vue';
 import CommandeView from '../views/CommandeView.vue';
 import CommanderView from '../views/CommanderView.vue';
-import ReservationsView from '../views/ReservationsView.vue';
 import ChefView from '../views/ChefView.vue';
 import OrderList from '../components/chef/OrderList.vue';
 import RecipeManager from '../components/chef/RecipeManager.vue';
-import TableDisponibiliteView from '../views/TableDisponibiliteView.vue'
+import ReservationView from '../views/ReservationView.vue'
 
 
 
@@ -32,17 +29,14 @@ const routes = [
   {path: '/admin/orders',name: 'admin-orders',component: OrderManagement, meta: { requiresAuth: true, role: 'ADMIN' }},
   {path: '/admin/statistics',name: 'admin-statistics',component: StatisticsManagement, meta: { requiresAuth: true, role: 'ADMIN' }},
   {path: '/user',name: 'user',component: UserView, meta: { requiresAuth: true, role: 'CLIENT' }},
-  {path: '/user/reservation',name: 'reservation',component: ReservationView, meta: { requiresAuth: true, role: 'CLIENT' }},
-  {path: '/user/reservation/choix',name: 'reservationChoix',component: ReservationChoiceView, meta: { requiresAuth: true, role: 'CLIENT' }},
   {path: '/user/feedback',name: 'feedback',component: FeedbackView, meta: { requiresAuth: true, role: 'CLIENT' }},
   {path: '/user/modifier',name: 'modifier',component: ModifyView, meta: { requiresAuth: true, role: 'CLIENT' }},
   {path: '/user/commandes',name: 'commandes',component: CommandeView, meta: { requiresAuth: true, role: 'CLIENT' }},
   {path: '/user/commander',name: 'commander',component: CommanderView, meta: { requiresAuth: true, role: 'CLIENT' }},
-  {path: '/user/reservations',name: 'mesReservations',component: ReservationsView, meta: { requiresAuth: true, role: 'CLIENT' }},
   {path: '/chef/',name: 'chef',component: ChefView, meta: { requiresAuth: true, role: 'CUISINIER' }},
   {path: '/chef/orders',name: 'ChefOrders',component: OrderList, meta: { requiresAuth: true, role: 'CUISINIER' }},
   {path: '/chef/recipes',name: 'chefRecipe',component: RecipeManager, meta: { requiresAuth: true, role: 'CUISINIER' }},
-  {path: '/user/tables-disponibles',name: 'tablesDisponibles',component: TableDisponibiliteView ,meta: { requiresAuth: true, role: 'CLIENT' }},
+  {path: '/user/reservation',name: 'reservation',component: ReservationView ,meta: { requiresAuth: true, role: 'CLIENT' }},
 ];
 
 const router = createRouter({
